+++ flag/templatetags/flag_tags.py
@@ -1,8 +1,7 @@
 from django import template
-from django.db.models import ObjectDoesNotExist
 
 from flag.forms import get_default_form
-from flag.views import get_next, get_confirm_url_for_object
+from flag.views import get_next, get_confirm_url_for_object, can_be_flagged_by
 from flag.models import FlaggedContent
 
 register = template.Library()
@@ -60,29 +59,7 @@ def flag_status(content_object, full=False):
     except:
         return None
 
-
-@register.filter
-def can_be_flagged_by(content_object, user):
-    """
-    This filter will return True if the given user can flag the given object.
-    We check that the user is authenticated, but also that the
-    LIMIT_SAME_OBJECT_FOR_USER is not raised
-    Usage: {% if some_object|can_by_flagged_by:request.user %}...{% endif %}
-    """
-    try:
-        if not (user and user.is_active and user.is_authenticated()):
-            return False
-        if not FlaggedContent.objects.model_can_be_flagged(content_object):
-            return False
-        try:
-            flagged_content = FlaggedContent.objects.get_for_object(
-                    content_object)
-            return flagged_content.can_be_flagged_by_user(user)
-        except ObjectDoesNotExist:
-            # no FlaggedContent, we know it canbe flagged
-            return True
-    except:
-        return False
+register.filter(can_be_flagged_by)
 
 
 @register.filter
